# Speed up creating Ruby AST