When creating the Ruby AST, we were previously allocating Location objects for every node and every inner location. Instead, this commit changes it to pack both the start offset and length into a single u64 and pass that into the nodes. Then, when the locations are requested via a reader method, we lazily allocate the Location objects.

This pretty drastically improves speed and memory. With this script:

```ruby
require "bundler/setup"
require "prism"

filepaths = Dir[File.expand_path("path/to/extremely/large/ruby/codebase/**/*.rb", __dir__)]
filepaths.each { |filepath| Prism.parse_file(filepath) }
puts `ps -o rss= -p #{$$}`
```

We get the following results. On `main`:

```
145152
ruby bin/bench  26.96s user 2.53s system 94% cpu 31.059 total

146016
ruby bin/bench  27.00s user 2.58s system 93% cpu 31.570 total

156048
ruby bin/bench  26.87s user 2.55s system 93% cpu 31.336 total
```

On `main` with `--yjit`:

```
166384
ruby --yjit bin/bench  17.27s user 2.55s system 91% cpu 21.622 total

164704
ruby --yjit bin/bench  17.31s user 2.53s system 92% cpu 21.495 total

154672
ruby --yjit bin/bench  17.63s user 2.72s system 91% cpu 22.302 total
```

On this branch:

```
132816
ruby bin/bench  15.06s user 2.54s system 90% cpu 19.538 total

127344
ruby bin/bench  15.05s user 2.45s system 91% cpu 19.056 total

129152
ruby bin/bench  15.13s user 2.49s system 91% cpu 19.202 total
```

On this branch with `--yjit`:

```
140688
ruby --yjit bin/bench  10.32s user 2.46s system 88% cpu 14.496 total

137888
ruby --yjit bin/bench  10.32s user 2.43s system 88% cpu 14.403 total

138704
ruby --yjit bin/bench  10.32s user 2.46s system 89% cpu 14.355 total
```