@eregon 2024-02-15
Great stuff!
It reminds of a similar change I did in https://github.com/ruby/prism/pull/928/commits/9c57958604d87eff3641c672bdbf6316349b291c but this is better as it packs both start & length in a single ivar.